Created general outline for iomanager class
