+++ assets/game.cfg
@@ -1,5 +1,5 @@
 #General config file
-[files]
+[configs]
   imageconfig = assets/images.cfg
   mapconfig = assets/maps.cfg
 [graphics]
